+++ homeassistant/components/homekit/type_media_players.py
@@ -1,7 +1,7 @@
 """Class to hold all media player accessories."""
 import logging
 
-from pyhap.const import CATEGORY_SWITCH, CATEGORY_TELEVISION
+from pyhap.const import CATEGORY_SWITCH
 
 from homeassistant.components.media_player import (
     ATTR_INPUT_SOURCE,
@@ -17,13 +17,6 @@ from homeassistant.components.media_player import (
     SUPPORT_VOLUME_SET,
     SUPPORT_VOLUME_STEP,
 )
-from homeassistant.components.remote import (
-    ATTR_ACTIVITY,
-    ATTR_ACTIVITY_LIST,
-    ATTR_CURRENT_ACTIVITY,
-    DOMAIN as REMOTE_DOMAIN,
-    SUPPORT_ACTIVITY,
-)
 from homeassistant.const import (
     ATTR_ENTITY_ID,
     ATTR_SUPPORTED_FEATURES,
@@ -38,7 +31,6 @@ from homeassistant.const import (
     SERVICE_VOLUME_SET,
     SERVICE_VOLUME_UP,
     STATE_OFF,
-    STATE_ON,
     STATE_PAUSED,
     STATE_PLAYING,
     STATE_STANDBY,
@@ -50,17 +42,9 @@ from .accessories import TYPES, HomeAccessory
 from .const import (
     ATTR_KEY_NAME,
     CHAR_ACTIVE,
-    CHAR_ACTIVE_IDENTIFIER,
-    CHAR_CONFIGURED_NAME,
-    CHAR_CURRENT_VISIBILITY_STATE,
-    CHAR_IDENTIFIER,
-    CHAR_INPUT_SOURCE_TYPE,
-    CHAR_IS_CONFIGURED,
     CHAR_MUTE,
     CHAR_NAME,
     CHAR_ON,
-    CHAR_REMOTE_KEY,
-    CHAR_SLEEP_DISCOVER_MODE,
     CHAR_VOLUME,
     CHAR_VOLUME_CONTROL_TYPE,
     CHAR_VOLUME_SELECTOR,
@@ -70,43 +54,15 @@ from .const import (
     FEATURE_PLAY_PAUSE,
     FEATURE_PLAY_STOP,
     FEATURE_TOGGLE_MUTE,
-    KEY_ARROW_DOWN,
-    KEY_ARROW_LEFT,
-    KEY_ARROW_RIGHT,
-    KEY_ARROW_UP,
-    KEY_BACK,
-    KEY_EXIT,
-    KEY_FAST_FORWARD,
-    KEY_INFORMATION,
-    KEY_NEXT_TRACK,
     KEY_PLAY_PAUSE,
-    KEY_PREVIOUS_TRACK,
-    KEY_REWIND,
-    KEY_SELECT,
-    SERV_INPUT_SOURCE,
     SERV_SWITCH,
-    SERV_TELEVISION,
     SERV_TELEVISION_SPEAKER,
 )
+from .type_remote import REMOTE_KEYS, RemoteInputSelectAccessory
 from .util import get_media_player_features
 
 _LOGGER = logging.getLogger(__name__)
 
-MEDIA_PLAYER_KEYS = {
-    0: KEY_REWIND,
-    1: KEY_FAST_FORWARD,
-    2: KEY_NEXT_TRACK,
-    3: KEY_PREVIOUS_TRACK,
-    4: KEY_ARROW_UP,
-    5: KEY_ARROW_DOWN,
-    6: KEY_ARROW_LEFT,
-    7: KEY_ARROW_RIGHT,
-    8: KEY_SELECT,
-    9: KEY_BACK,
-    10: KEY_EXIT,
-    11: KEY_PLAY_PAUSE,
-    15: KEY_INFORMATION,
-}
 
 # Names may not contain special characters
 # or emjoi (/ is a special character for Apple)
@@ -257,132 +213,8 @@ class MediaPlayer(HomeAccessory):
                 self.chars[FEATURE_TOGGLE_MUTE].set_value(current_state)
 
 
-class RemoteBase(HomeAccessory):
-    """Generate a Television accessory."""
-
-    def __init__(self, required_feature, source_key, source_list_key, *args):
-        """Initialize a Remote accessory object."""
-        super().__init__(*args, category=CATEGORY_TELEVISION)
-        state = self.hass.states.get(self.entity_id)
-        features = state.attributes.get(ATTR_SUPPORTED_FEATURES, 0)
-
-        self.source_key = source_key
-        self.sources = []
-        self.support_select_source = False
-        if features & required_feature:
-            self.sources = state.attributes.get(source_list_key, [])
-            if self.sources:
-                self.support_select_source = True
-
-        self.chars_tv = [CHAR_REMOTE_KEY]
-
-        serv_tv = self.serv_tv = self.add_preload_service(
-            SERV_TELEVISION, self.chars_tv
-        )
-        self.char_remote_key = self.serv_tv.configure_char(
-            CHAR_REMOTE_KEY, setter_callback=self.set_remote_key
-        )
-        self.set_primary_service(serv_tv)
-        serv_tv.configure_char(CHAR_CONFIGURED_NAME, value=self.display_name)
-        serv_tv.configure_char(CHAR_SLEEP_DISCOVER_MODE, value=True)
-        self.char_active = serv_tv.configure_char(
-            CHAR_ACTIVE, setter_callback=self.set_on_off
-        )
-
-        if self.support_select_source:
-            self.char_input_source = serv_tv.configure_char(
-                CHAR_ACTIVE_IDENTIFIER, setter_callback=self.set_input_source
-            )
-            for index, source in enumerate(self.sources):
-                serv_input = self.add_preload_service(
-                    SERV_INPUT_SOURCE, [CHAR_IDENTIFIER, CHAR_NAME]
-                )
-                serv_tv.add_linked_service(serv_input)
-                serv_input.configure_char(CHAR_CONFIGURED_NAME, value=source)
-                serv_input.configure_char(CHAR_NAME, value=source)
-                serv_input.configure_char(CHAR_IDENTIFIER, value=index)
-                serv_input.configure_char(CHAR_IS_CONFIGURED, value=True)
-                input_type = 3 if "hdmi" in source.lower() else 0
-                serv_input.configure_char(CHAR_INPUT_SOURCE_TYPE, value=input_type)
-                serv_input.configure_char(CHAR_CURRENT_VISIBILITY_STATE, value=False)
-                _LOGGER.debug("%s: Added source %s", self.entity_id, source)
-
-    @callback
-    def _async_update_input_state(self, hk_state, new_state):
-        """Update input state after state changed."""
-        # Set active input
-        if self.support_select_source and self.sources:
-            source_name = new_state.attributes.get(self.source_key)
-            _LOGGER.debug("%s: Set current input to %s", self.entity_id, source_name)
-            if source_name in self.sources:
-                index = self.sources.index(source_name)
-                if self.char_input_source.value != index:
-                    self.char_input_source.set_value(index)
-            elif hk_state:
-                _LOGGER.warning(
-                    "%s: Sources out of sync. Restart Home Assistant",
-                    self.entity_id,
-                )
-                if self.char_input_source.value != 0:
-                    self.char_input_source.set_value(0)
-
-
-@TYPES.register("TelevisionRemote")
-class TelevisionRemote(RemoteBase):
-    """Generate a Television Remote accessory."""
-
-    def __init__(self, *args):
-        """Initialize a Television Remote accessory object."""
-        super().__init__(
-            SUPPORT_ACTIVITY,
-            ATTR_CURRENT_ACTIVITY,
-            ATTR_ACTIVITY_LIST,
-            *args,
-        )
-        state = self.hass.states.get(self.entity_id)
-        self.async_update_state(state)
-
-    def set_on_off(self, value):
-        """Move switch state to value if call came from HomeKit."""
-        _LOGGER.debug('%s: Set switch state for "on_off" to %s', self.entity_id, value)
-        service = SERVICE_TURN_ON if value else SERVICE_TURN_OFF
-        params = {ATTR_ENTITY_ID: self.entity_id}
-        self.async_call_service(REMOTE_DOMAIN, service, params)
-
-    def set_input_source(self, value):
-        """Send input set value if call came from HomeKit."""
-        _LOGGER.debug("%s: Set current input to %s", self.entity_id, value)
-        source = self.sources[value]
-        params = {ATTR_ENTITY_ID: self.entity_id, ATTR_ACTIVITY: source}
-        self.async_call_service(REMOTE_DOMAIN, SERVICE_TURN_ON, params)
-
-    def set_remote_key(self, value):
-        """Send remote key value if call came from HomeKit."""
-        _LOGGER.debug("%s: Set remote key to %s", self.entity_id, value)
-        key_name = MEDIA_PLAYER_KEYS.get(value)
-        if key_name is None:
-            _LOGGER.warning("%s: Unhandled key press for %s", self.entity_id, value)
-            return
-        self.hass.bus.async_fire(
-            EVENT_HOMEKIT_TV_REMOTE_KEY_PRESSED,
-            {ATTR_KEY_NAME: key_name, ATTR_ENTITY_ID: self.entity_id},
-        )
-
-    @callback
-    def async_update_state(self, new_state):
-        """Update Television remote state after state changed."""
-        current_state = new_state.state
-        # Power state remote
-        hk_state = 1 if current_state == STATE_ON else 0
-        _LOGGER.debug("%s: Set current active state to %s", self.entity_id, hk_state)
-        if self.char_active.value != hk_state:
-            self.char_active.set_value(hk_state)
-
-        self._async_update_input_state(hk_state, new_state)
-
-
 @TYPES.register("TelevisionMediaPlayer")
-class TelevisionMediaPlayer(RemoteBase):
+class TelevisionMediaPlayer(RemoteInputSelectAccessory):
     """Generate a Television Media Player accessory."""
 
     def __init__(self, *args):
@@ -478,7 +310,7 @@ class TelevisionMediaPlayer(RemoteBase):
     def set_remote_key(self, value):
         """Send remote key value if call came from HomeKit."""
         _LOGGER.debug("%s: Set remote key to %s", self.entity_id, value)
-        key_name = MEDIA_PLAYER_KEYS.get(value)
+        key_name = REMOTE_KEYS.get(value)
         if key_name is None:
             _LOGGER.warning("%s: Unhandled key press for %s", self.entity_id, value)
             return
